Break interpreter and model checker into own files and fix segfault
